+++ lib/main.js
@@ -250,6 +250,7 @@ const Hydrogen = {
     store.dispose();
   },
 
+  /*-------------- Service Providers --------------*/
   provideHydrogen() {
     if (!this.hydrogenProvider) {
       this.hydrogenProvider = new HydrogenProvider(this);
@@ -258,17 +259,20 @@ const Hydrogen = {
     return this.hydrogenProvider;
   },
 
+  provideAutocompleteResults() {
+    return services.provided.autocomplete.provideAutocompleteResults(store);
+  },
+  /*-----------------------------------------------*/
+
+  /*-------------- Service Consumers --------------*/
   consumeStatusBar(statusBar: atom$StatusBar) {
-    return services.statusBar.addStatusBar(
+    return services.consumed.statusBar.addStatusBar(
       store,
       statusBar,
       this.handleKernelCommand.bind(this)
     );
   },
-
-  provideAutocompleteResults() {
-    return services.autocomplete.provideAutocompleteResults(store);
-  },
+  /*-----------------------------------------------*/
 
   connectToExistingKernel() {
     if (!this.existingKernelPicker) {
