--- lib/services/autocomplete/index.js
@@ -1,4 +1,0 @@
-/* @flow */
-import * as autocompleteProvider from "./autocomplete-provider";
-
-export default { ...autocompleteProvider };
