+++ lib/services/consumed/index.js
@@ -1,0 +1,4 @@
+/* @flow */
+import * as statusBar from "./status-bar";
+
+export default { statusBar };
