+++ lib/services/consumed/status-bar/component.js
@@ -3,8 +3,8 @@
 import React from "react";
 import { observer } from "mobx-react";
 
-import type Kernel from "../../kernel";
-import type { Store } from "../../store";
+import type Kernel from "../../../kernel";
+import type { Store } from "../../../store";
 
 type Props = {
   store: Store,
@@ -12,7 +12,7 @@ type Props = {
 };
 
 @observer
-class StatusBar extends React.Component<Props> {
+export class StatusBar extends React.Component<Props> {
   render() {
     const { kernel, configMapping } = this.props.store;
     if (!kernel || configMapping.get("Hydrogen.statusBarDisable")) return null;
@@ -23,5 +23,3 @@ class StatusBar extends React.Component<Props> {
     );
   }
 }
-
-export default StatusBar;
