+++ lib/services/consumed/status-bar/index.js
@@ -1,0 +1,6 @@
+/* @flow */
+import * as component from "./component";
+import * as signalListView from "./signal-list-view";
+import * as statusBar from "./status-bar";
+
+export default { ...component, ...signalListView, ...statusBar };
