+++ lib/services/consumed/status-bar/signal-list-view.js
@@ -2,8 +2,8 @@
 
 import SelectListView from "atom-select-list";
 
-import WSKernel from "../../ws-kernel";
-import { log } from "../../utils";
+import WSKernel from "../../../ws-kernel";
+import { log } from "../../../utils";
 
 const basicCommands = [
   { name: "Interrupt", value: "interrupt-kernel" },
@@ -16,7 +16,7 @@ const wsKernelCommands = [
   { name: "Disconnect from", value: "disconnect-kernel" }
 ];
 
-export default class SignalListView {
+export class SignalListView {
   onConfirmed: ?(command: { command: string }) => void;
   panel: ?atom$Panel;
   previouslyFocusedElement: ?HTMLElement;
