+++ lib/services/consumed/status-bar/status-bar.js
@@ -3,13 +3,13 @@
 import React from "react";
 import { Disposable } from "atom";
 
-import StatusBar from "./status-bar";
-import SignalListView from "./signal-list-view";
+import { StatusBar } from "./component";
+import { SignalListView } from "./signal-list-view";
 
-import { reactFactory } from "../../utils";
+import { reactFactory } from "../../../utils";
 
-import type { Store } from "../../store";
-import type Kernel from "../../kernel";
+import type { Store } from "../../../store";
+import type Kernel from "../../../kernel";
 
 export function addStatusBar(
   store: Store,
