+++ lib/services/index.js
@@ -1,5 +1,5 @@
 /* @flow */
-import * as statusBar from "./status-bar";
-import autocomplete from "./autocomplete";
+import consumed from "./consumed";
+import provided from "./provided";
 
-export default { statusBar, autocomplete };
+export default { consumed, provided };
