+++ lib/services/provided/index.js
@@ -1,0 +1,4 @@
+/* @flow */
+import * as autocomplete from "./autocomplete";
+
+export default { autocomplete };
